// src/models/employee.ts
export interface IEmployee {
  name: string;
  phone: string;
  email: string;
  dateOfBirth: string;
  gender: number;
  address: string;
  id: number;
  code: string;
  username: string;
  avatar: string;
  position: string;
  startTime: string;
  endTime: string;
  status: number;
  createdAt: string;
  updatedAt: string;
  isRemoved: boolean;
  branchId: number;
  city: string;
  district: string;
  ward: string;
}